+++ Flow Graph Library/Inspection Flow Graphs/File/signal_envelope.py
@@ -1,0 +1,249 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+#
+# SPDX-License-Identifier: GPL-3.0
+#
+# GNU Radio Python Flow Graph
+# Title: Signal Envelope
+# GNU Radio version: 3.8.5.0
+
+from distutils.version import StrictVersion
+
+if __name__ == '__main__':
+    import ctypes
+    import sys
+    if sys.platform.startswith('linux'):
+        try:
+            x11 = ctypes.cdll.LoadLibrary('libX11.so')
+            x11.XInitThreads()
+        except:
+            print("Warning: failed to XInitThreads()")
+
+from PyQt5 import Qt
+from PyQt5.QtCore import QObject, pyqtSlot
+from gnuradio import qtgui
+from gnuradio.filter import firdes
+import sip
+from gnuradio import blocks
+import pmt
+from gnuradio import gr
+import sys
+import signal
+from argparse import ArgumentParser
+from gnuradio.eng_arg import eng_float, intx
+from gnuradio import eng_notation
+
+from gnuradio import qtgui
+
+class signal_envelope(gr.top_block, Qt.QWidget):
+
+    def __init__(self, filepath='', sample_rate='1e6'):
+        gr.top_block.__init__(self, "Signal Envelope")
+        Qt.QWidget.__init__(self)
+        self.setWindowTitle("Signal Envelope")
+        qtgui.util.check_set_qss()
+        try:
+            self.setWindowIcon(Qt.QIcon.fromTheme('gnuradio-grc'))
+        except:
+            pass
+        self.top_scroll_layout = Qt.QVBoxLayout()
+        self.setLayout(self.top_scroll_layout)
+        self.top_scroll = Qt.QScrollArea()
+        self.top_scroll.setFrameStyle(Qt.QFrame.NoFrame)
+        self.top_scroll_layout.addWidget(self.top_scroll)
+        self.top_scroll.setWidgetResizable(True)
+        self.top_widget = Qt.QWidget()
+        self.top_scroll.setWidget(self.top_widget)
+        self.top_layout = Qt.QVBoxLayout(self.top_widget)
+        self.top_grid_layout = Qt.QGridLayout()
+        self.top_layout.addLayout(self.top_grid_layout)
+
+        self.settings = Qt.QSettings("GNU Radio", "signal_envelope")
+
+        try:
+            if StrictVersion(Qt.qVersion()) < StrictVersion("5.0.0"):
+                self.restoreGeometry(self.settings.value("geometry").toByteArray())
+            else:
+                self.restoreGeometry(self.settings.value("geometry"))
+        except:
+            pass
+
+        ##################################################
+        # Parameters
+        ##################################################
+        self.filepath = filepath
+        self.sample_rate = sample_rate
+
+        ##################################################
+        # Variables
+        ##################################################
+        self.decimation = decimation = 1
+
+        ##################################################
+        # Blocks
+        ##################################################
+        # Create the options list
+        self._decimation_options = [1, 10, 100, 1000]
+        # Create the labels list
+        self._decimation_labels = ['1', '10', '100', '1000']
+        # Create the combo box
+        self._decimation_tool_bar = Qt.QToolBar(self)
+        self._decimation_tool_bar.addWidget(Qt.QLabel('  Keep 1 in N' + ": "))
+        self._decimation_combo_box = Qt.QComboBox()
+        self._decimation_tool_bar.addWidget(self._decimation_combo_box)
+        for _label in self._decimation_labels: self._decimation_combo_box.addItem(_label)
+        self._decimation_callback = lambda i: Qt.QMetaObject.invokeMethod(self._decimation_combo_box, "setCurrentIndex", Qt.Q_ARG("int", self._decimation_options.index(i)))
+        self._decimation_callback(self.decimation)
+        self._decimation_combo_box.currentIndexChanged.connect(
+            lambda i: self.set_decimation(self._decimation_options[i]))
+        # Create the radio buttons
+        self.top_grid_layout.addWidget(self._decimation_tool_bar, 0, 2, 1, 1)
+        for r in range(0, 1):
+            self.top_grid_layout.setRowStretch(r, 1)
+        for c in range(2, 3):
+            self.top_grid_layout.setColumnStretch(c, 1)
+        self.qtgui_time_sink_x_0 = qtgui.time_sink_f(
+            100000, #size
+            float(sample_rate)/decimation, #samp_rate
+            "", #name
+            1 #number of inputs
+        )
+        self.qtgui_time_sink_x_0.set_update_time(0.1)
+        self.qtgui_time_sink_x_0.set_y_axis(-.1, 1)
+
+        self.qtgui_time_sink_x_0.set_y_label('Amplitude', "")
+
+        self.qtgui_time_sink_x_0.enable_tags(True)
+        self.qtgui_time_sink_x_0.set_trigger_mode(qtgui.TRIG_MODE_FREE, qtgui.TRIG_SLOPE_POS, 0.005, 0, 0, "")
+        self.qtgui_time_sink_x_0.enable_autoscale(False)
+        self.qtgui_time_sink_x_0.enable_grid(True)
+        self.qtgui_time_sink_x_0.enable_axis_labels(True)
+        self.qtgui_time_sink_x_0.enable_control_panel(False)
+        self.qtgui_time_sink_x_0.enable_stem_plot(False)
+
+
+        labels = ['', '', '', '', '',
+            '', '', '', '', '']
+        widths = [1, 1, 1, 1, 1,
+            1, 1, 1, 1, 1]
+        colors = ['blue', 'red', 'green', 'black', 'cyan',
+            'magenta', 'yellow', 'dark red', 'dark green', 'dark blue']
+        alphas = [1.0, 1.0, 1.0, 1.0, 1.0,
+            1.0, 1.0, 1.0, 1.0, 1.0]
+        styles = [1, 1, 1, 1, 1,
+            1, 1, 1, 1, 1]
+        markers = [-1, -1, -1, -1, -1,
+            -1, -1, -1, -1, -1]
+
+
+        for i in range(1):
+            if len(labels[i]) == 0:
+                self.qtgui_time_sink_x_0.set_line_label(i, "Data {0}".format(i))
+            else:
+                self.qtgui_time_sink_x_0.set_line_label(i, labels[i])
+            self.qtgui_time_sink_x_0.set_line_width(i, widths[i])
+            self.qtgui_time_sink_x_0.set_line_color(i, colors[i])
+            self.qtgui_time_sink_x_0.set_line_style(i, styles[i])
+            self.qtgui_time_sink_x_0.set_line_marker(i, markers[i])
+            self.qtgui_time_sink_x_0.set_line_alpha(i, alphas[i])
+
+        self._qtgui_time_sink_x_0_win = sip.wrapinstance(self.qtgui_time_sink_x_0.pyqwidget(), Qt.QWidget)
+        self.top_grid_layout.addWidget(self._qtgui_time_sink_x_0_win, 3, 0, 20, 4)
+        for r in range(3, 23):
+            self.top_grid_layout.setRowStretch(r, 1)
+        for c in range(0, 4):
+            self.top_grid_layout.setColumnStretch(c, 1)
+        self.blocks_throttle_0 = blocks.throttle(gr.sizeof_gr_complex*1, float(sample_rate),True)
+        self.blocks_keep_one_in_n_0 = blocks.keep_one_in_n(gr.sizeof_gr_complex*1, decimation)
+        self.blocks_file_source_0 = blocks.file_source(gr.sizeof_gr_complex*1, filepath, True, 0, 0)
+        self.blocks_file_source_0.set_begin_tag(pmt.PMT_NIL)
+        self.blocks_complex_to_mag_squared_0 = blocks.complex_to_mag_squared(1)
+
+
+        ##################################################
+        # Connections
+        ##################################################
+        self.connect((self.blocks_complex_to_mag_squared_0, 0), (self.qtgui_time_sink_x_0, 0))
+        self.connect((self.blocks_file_source_0, 0), (self.blocks_throttle_0, 0))
+        self.connect((self.blocks_keep_one_in_n_0, 0), (self.blocks_complex_to_mag_squared_0, 0))
+        self.connect((self.blocks_throttle_0, 0), (self.blocks_keep_one_in_n_0, 0))
+
+
+    def closeEvent(self, event):
+        self.settings = Qt.QSettings("GNU Radio", "signal_envelope")
+        self.settings.setValue("geometry", self.saveGeometry())
+        event.accept()
+
+    def get_filepath(self):
+        return self.filepath
+
+    def set_filepath(self, filepath):
+        self.filepath = filepath
+        self.blocks_file_source_0.open(self.filepath, True)
+
+    def get_sample_rate(self):
+        return self.sample_rate
+
+    def set_sample_rate(self, sample_rate):
+        self.sample_rate = sample_rate
+        self.blocks_throttle_0.set_sample_rate(float(self.sample_rate))
+        self.qtgui_time_sink_x_0.set_samp_rate(float(self.sample_rate)/self.decimation)
+
+    def get_decimation(self):
+        return self.decimation
+
+    def set_decimation(self, decimation):
+        self.decimation = decimation
+        self._decimation_callback(self.decimation)
+        self.blocks_keep_one_in_n_0.set_n(self.decimation)
+        self.qtgui_time_sink_x_0.set_samp_rate(float(self.sample_rate)/self.decimation)
+
+
+
+
+def argument_parser():
+    parser = ArgumentParser()
+    parser.add_argument(
+        "--filepath", dest="filepath", type=str, default='',
+        help="Set filepath [default=%(default)r]")
+    parser.add_argument(
+        "--sample-rate", dest="sample_rate", type=str, default='1e6',
+        help="Set 1e6 [default=%(default)r]")
+    return parser
+
+
+def main(top_block_cls=signal_envelope, options=None):
+    if options is None:
+        options = argument_parser().parse_args()
+
+    if StrictVersion("4.5.0") <= StrictVersion(Qt.qVersion()) < StrictVersion("5.0.0"):
+        style = gr.prefs().get_string('qtgui', 'style', 'raster')
+        Qt.QApplication.setGraphicsSystem(style)
+    qapp = Qt.QApplication(sys.argv)
+
+    tb = top_block_cls(filepath=options.filepath, sample_rate=options.sample_rate)
+
+    tb.start()
+
+    tb.show()
+
+    def sig_handler(sig=None, frame=None):
+        Qt.QApplication.quit()
+
+    signal.signal(signal.SIGINT, sig_handler)
+    signal.signal(signal.SIGTERM, sig_handler)
+
+    timer = Qt.QTimer()
+    timer.start(500)
+    timer.timeout.connect(lambda: None)
+
+    def quitting():
+        tb.stop()
+        tb.wait()
+
+    qapp.aboutToQuit.connect(quitting)
+    qapp.exec_()
+
+if __name__ == '__main__':
+    main()